requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
filelock==3.8.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.4
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
virtualenv==20.16.7
XlsxWriter==3.0.3
yarg==0.1.9
zipp==3.11.0
